+++ CHANGELOG.md
@@ -28,3 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added OnDestroy method to addressables manager
+
+## [1.0.5] - 2023-02-25
+### Changed
+- changed header text to better reflect fields
+- made fields public
+
+### Added
+- Added a MiniGameDataMode which can be set to indicate where to get the data from